Refactor the meta select into it’s own component
